% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cash.4.life}
\alias{cash.4.life}
\title{CASH4LIFE}
\format{A data frame with 34 rows and 7 variables:
\describe{
  \item{date}{date of draw}
  \item{cashball}{cash ball result}
  \item{N1}{1st number in order}
  \item{N2}{2nd number in order}
  \item{N3}{3rd number in order}
  \item{N4}{4th number in order}
  \item{N5}{5th number in order}
}}
\source{
\url{https://www.valottery.com}
}
\usage{
cash.4.life
}
\description{
Historical data for the CASH4LIFE game. Game play: Pick five different numbers from 1 through 60;
then select one Cash Ball number from 1 through 4.
}
\examples{
## Check numbers drawn are uniformly distributed
x <- qunif(ppoints(nrow(cash.4.life)*5),1,60)
y <- sort(unlist(cash.4.life[,3:7]))
qqplot(x,y)
}
\keyword{datasets}

