% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cash.5.1xday}
\alias{cash.5.1xday}
\title{Cash 5 (once daily)}
\format{A data frame with 1187 rows and 6 variables:
\describe{
  \item{date}{date of draw}
  \item{N1}{1st number in order}
  \item{N2}{2nd number in order}
  \item{N3}{3rd number in order}
  \item{N4}{4th number in order}
  \item{N5}{5th number in order}
}}
\source{
\url{https://www.valottery.com}
}
\usage{
cash.5.1xday
}
\description{
Historical data for the Cash 5 once daily game. Game Play: Pick five
numbers from 1 through 34. Note: On April 11, 1999, Cash 5 switched
to twice daily drawings.
}
\keyword{datasets}

