% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mega.mill.1}
\alias{mega.mill.1}
\title{Mega Millions (Before 10/22/13)}
\format{A data frame with 1,713 rows and 7 variables:
\describe{
  \item{date}{date of draw}
  \item{megaball}{megaball result}
  \item{N1}{1st number in order}
  \item{N2}{2nd number in order}
  \item{N3}{3rd number in order}
  \item{N4}{4th number in order}
  \item{N5}{5th number in order}
}}
\source{
\url{https://www.valottery.com}
}
\usage{
mega.mill.1
}
\description{
Historical data for the Mega Millions game. On October 22, 2013, the format changed
from 5/56 + 1/46 to the current 5/75 + 1/15 format. Game play: Pick five different
numbers from 1 through 75; then select one Mega Ball number from 1 through 15.
}
\keyword{datasets}

