% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pick.3.1xday}
\alias{pick.3.1xday}
\title{Pick 3 (once daily)}
\format{A data frame with 1777 rows and 4 variables:
\describe{
  \item{date}{date of draw}
  \item{N1}{1st number in order}
  \item{N2}{2nd number in order}
  \item{N3}{3rd number in order}
}}
\source{
\url{https://www.valottery.com}
}
\usage{
pick.3.1xday
}
\description{
Historical data for the Pick 3 once daily game. On January 30, 1995,
Pick 3 switched to twice daily drawings.
}
\keyword{datasets}

