% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pick.3.2xday}
\alias{pick.3.2xday}
\title{Pick 3 (twice daily)}
\format{A data frame with 13,790 rows and 5 variables:
\describe{
  \item{date}{date of draw}
  \item{time}{time of drawing: day, night1 or night2 (Note: two nightly drawings were held
  on 10/30/08 and 11/09/08)}
  \item{N1}{1st digit}
  \item{N2}{2nd digit}
  \item{N3}{3rd digit}
}}
\source{
\url{https://www.valottery.com}
}
\usage{
pick.3.2xday
}
\description{
Historical data for the Pick 3 twice daily game. Game Play:
Pick a three digit number from 000 through 999. Note: On January 30, 1995,
Pick 3 switched to twice daily drawings.
}
\keyword{datasets}

