% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pick.4.2xday}
\alias{pick.4.2xday}
\title{Pick 4 (twice daily)}
\format{A data frame with 13,788 rows and 6 variables:
\describe{
  \item{date}{date of draw}
  \item{time}{time of drawing: day or night}
  \item{N1}{1st digit}
  \item{N2}{2nd digit}
  \item{N3}{3rd digit}
  \item{N4}{4th digit}
}}
\source{
\url{https://www.valottery.com}
}
\usage{
pick.4.2xday
}
\description{
Historical data for the Pick 4 twice daily game. Game play: Pick
a four digit number from 0000 through 9999. Note: On January 30, 1995,
Pick 4 switched to twice daily drawings.
}
\keyword{datasets}

