% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{power.ball}
\alias{power.ball}
\title{Power Ball}
\format{A data frame with 582 rows and 7 variables:
\describe{
  \item{date}{date of draw}
  \item{powerball}{powerball result}
  \item{N1}{1st number in order}
  \item{N2}{2nd number in order}
  \item{N3}{3rd number in order}
  \item{N4}{4th number in order}
  \item{N5}{5th number in order}
}}
\source{
\url{https://www.valottery.com}
}
\usage{
power.ball
}
\description{
Historical data for the Power Ball game. Game play: Pick
five different numbers from 1 through 59; then select one
Powerball number from 1 through 35.
}
\examples{
## According to game rules, the powerball is numbered 1 - 35,
## but apparently there were times when it went up to 39
i <- power.ball$powerball > 35
any(i)
sum(i)
power.ball$powerball[i]
}
\keyword{datasets}

