% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pick.3.1xday}
\alias{pick.3.1xday}
\title{Pick 3 (once daily)}
\format{A data frame with 1,777 rows and 4 variables:
\describe{
  \item{date}{date of draw}
  \item{N1}{1st digit}
  \item{N2}{2nd digit}
  \item{N3}{3rd digit}
}}
\source{
\url{https://www.valottery.com}
}
\usage{
pick.3.1xday
}
\description{
Historical data for the Pick 3 once daily game. Game Play:
Pick a three digit number from 000 through 999. Note: On January 30, 1995,
Pick 3 switched to twice daily drawings.
}
\examples{
lapply(pick.3.1xday[,-1],function(x)round(prop.table(table(x)),2))
}
\keyword{datasets}

